games-emulation/eden: don't attempt to double-remove bundled externals

+++ games-emulation/eden/eden-0.2.0.ebuild
@@ -129,8 +129,10 @@ src_prepare() {
 		fi
 	done
 
-	einfo "removing sources: ${remove[*]}"
-	rm -r "${remove[@]}" || die
+	if (( ${#remove[@]} > 0 )); then
+		einfo "removing sources: ${remove[*]}"
+		rm -r "${remove[@]}" || die
+	fi
 
 	cmake_src_prepare
 }
